initial add

+++ src/User.java
@@ -11,8 +11,8 @@ public class User {
 	static final String DEFAULT_NAME = "<NO_NAME>";
 	String name;
 	static final int DEFAULT_NUM_MOVIES = 15;
-	double[] ratings = new double [DEFAULT_NUM_MOVIES];
-	double[] avgRatingsMinusMean = new double [DEFAULT_NUM_MOVIES];
+	int[] ratings = new int [DEFAULT_NUM_MOVIES];
+	int[] avgRatingsMinusMean = new int [DEFAULT_NUM_MOVIES];
 	boolean[] hasRatedMovie = new boolean [DEFAULT_NUM_MOVIES];
 	
 	/**
@@ -23,7 +23,7 @@ public class User {
 	{
 		this(DEFAULT_NAME);
 	}
-	 
+	
 	/**
 	 * One parameterized constructor
 	 * @param name
@@ -34,7 +34,7 @@ public class User {
 		this.SetName(name);
 		for(int i=0; i < DEFAULT_NUM_MOVIES; ++i)
 		{
-			this.ratings[i] = 0.0;
+			this.ratings[i] = 0;
 		}
 		CheckHasRatedMovie();
 		CalculateAvgRatingsMinusMean();
@@ -45,7 +45,7 @@ public class User {
 	 * @param name is passed as an argument to SetName 
 	 * @param ratings is passed as an argument to initialize the ratings
 	 */
-	public User(String name, double[] ratings)
+	public User(String name, int[] ratings)
 	{
 		this.SetName(name);
 		for(int i=0; i < DEFAULT_NUM_MOVIES; ++i)
@@ -56,24 +56,32 @@ public class User {
 		CalculateAvgRatingsMinusMean();
 	}
 	
+	/**
+	 * 
+	 */
 	public void CalculateAvgRatingsMinusMean()
 	{
-		double average = this.GetAverageRatings();
+		int average = this.GetAverageRatings();
 		
 		for(int i=0; i < User.DEFAULT_NUM_MOVIES; i++)
 		{
-			if(this.ratings[i] != 0.0)
+			if(this.ratings[i] != 0)
 			{
 				this.avgRatingsMinusMean[i] = this.ratings[i] - average;
 			}
 			else
 			{
-				this.avgRatingsMinusMean[i] = 0.0;
+				this.avgRatingsMinusMean[i] = 0;
 			}
 		}
 	}
 
-	public void AddRatingToMovie(int loc, double rating)
+	/**
+	 * 
+	 * @param loc
+	 * @param rating
+	 */
+	public void AddRatingToMovie(int loc, int rating)
 	{
 		this.ratings[loc] = rating;
 		CheckHasRatedMovie();
@@ -84,9 +92,9 @@ public class User {
 	 * 
 	 * @return
 	 */
-	public double GetAverageRatings()
+	public int GetAverageRatings()
 	{
-		double sum = 0.0;
+		int sum = 0;
 		int count = 0;
 		
 		for(int i=0; i < User.DEFAULT_NUM_MOVIES; i++)
@@ -100,7 +108,7 @@ public class User {
 		
 		if(count == 0)
 		{
-			return 0.0;
+			return 0;
 		}
 		
 		return sum / count;
@@ -128,20 +136,26 @@ public class User {
 	 * Mutator
 	 * @param ratings is the values rated for each movie
 	 */
-	public void SetRatings(double[] ratings)
+	public void SetRatings(int[] ratings)
 	{
 		this.ratings = ratings;
 		CheckHasRatedMovie();
 	}
 	
+	/**
+	 * 
+	 */
 	public void CheckHasRatedMovie()
 	{
 		for(int i=0; i<DEFAULT_NUM_MOVIES; i++)
 		{
-			this.hasRatedMovie[i] = (ratings[i] != 0.0) ? true : false;
+			this.hasRatedMovie[i] = (ratings[i] != 0) ? true : false;
 		}
 	}
 	
+	/**
+	 * 
+	 */
 	public void PrintRatings()
 	{
 		for(int i=0; i< DEFAULT_NUM_MOVIES; i++)
@@ -154,16 +168,24 @@ public class User {
 	 * Accessor
 	 * @return the ratings array which is the values rated for each movie
 	 */
-	public double[] GetRatings()
+	public int[] GetRatings()
 	{
 		return this.ratings;
 	}
 	
-	public double[] GetAvgRatingsMinusMean()
+	/**
+	 * 
+	 * @return
+	 */
+	public int[] GetAvgRatingsMinusMean()
 	{
 		return this.avgRatingsMinusMean;
 	}
 	
+	/**
+	 * 
+	 * @return
+	 */
 	public boolean[] GetHasRatedMovie()
 	{
 		return this.hasRatedMovie;
